+++ src/components/TodoList/TodoList.tsx
@@ -1,11 +1,11 @@
-import React from 'react';
-import { Todo } from '../../type/todo';
+import { FC } from 'react';
 import { TodoInfo } from '../TodoInfo';
+import { Todo } from '../../types/todo';
 interface Props {
   todos: Todo[];
 }
 
-export const TodoList: React.FC<Props> = ({ todos }) => (
+export const TodoList: FC<Props> = ({ todos }) => (
   <section className="TodoList">
     {todos.map(todo => (
       <TodoInfo key={todo.id} todo={todo} />
